5. List Slicing
Sometimes, you only want to access a portion of a list.

letters = ['a', 'b', 'c', 'd', 'e']
slice = letters[1:3]
print slice
print letters
In the above example, we first create a list called letters.
Then, we take a subsection and store it in the slice list. We start at the index before the colon and continue up to but not including the index after the colon.
Next, we print out ['b', 'c']. Remember that we start counting indices from 0 and that we stopped before index 3.
Finally, we print out ['a', 'b', 'c', 'd', 'e'], just to show that we did not modify the original letters list.
Instructions
On line 4, create a list called middle containing only the two middle items from suitcase.
On line 5, create a list called last made up only of the last two items from suitcase.